Outsource plotting.
