+++ pytrms/__init__.py
@@ -3,6 +3,10 @@ _version = '0.2.0'
 
 __all__ = []
 
+from .plotting import plot_marker
+
+__all__ += ['plot_marker']
+
 
 def load(path):
     '''Open a datafile for post-analysis or batch processing.
@@ -28,6 +32,7 @@ def connect(host='localhost', port=8002):
     from .clients.ioniclient import IoniClient
     from .tracebuffer import TraceBuffer
     from .instrument import Instrument
+    from .helpers import PTRConnectionError
 
     global _client
     global _buffer
@@ -37,30 +42,11 @@ def connect(host='localhost', port=8002):
     if _buffer is None:
         _buffer = TraceBuffer(_client)
 
-    return Instrument(_client, _buffer)
+    try:
+        inst = Instrument(_client, _buffer)
+    except PTRConnectionError as exc:
+        print(exc)
+        return None
 
-
-def plot_marker(signal, marker, **kwargs):
-    '''Plot a `signal` and fill the regions where `marker=True`.
-
-    Returns a tuple of `figure, axis`.
-    '''
-    import matplotlib.pyplot as plt
-
-    fig, ax = plt.subplots()
-    if hasattr(signal, 'plot'):
-        subplot = signal.plot(ax=ax)
-        line, *_ = subplot.get_lines()
-    else:
-        line, = ax.plot(signal)
-
-    x_ = line.get_xdata()
-    lo, hi = ax.get_ylim()
-    ax.fill_between(x_, lo, hi, where=marker, color='orange')
-
-    ax.grid(visible=True)
-    if hasattr(signal, 'name'):
-        ax.set_title(signal.name)
-
-    return fig, ax
+    return inst
